requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
yt_dlp==2024.7.9
discord.py==2.4.0